Anpassungen, neues Element

+++ web/iframe.php
@@ -18,18 +18,25 @@
             min-width: 100%;
             background-color: #eee;
             border: 5px solid red;
+            display: flex;
+            flex-flow: row wrap;
+            justify-content: space-around;
+            align-content: center;
+            align-items: center;
         }
 
         * {
             box-sizing: border-box
         }
     </style>
+    <script src="/files/base/layout/_vendor/node_modules/js-cookie/dist/js.cookie.min.js"></script>
+
+
 </head>
 <body>
 
 
-<div
-        style="height: 100%; min-width: 100%; padding: 15px; text-align: center; display: flex; flex-flow: row wrap; justify-content: space-around; align-content: center; align-items: center;">
+<div style="height: 100%; min-width: 100%; padding: 15px; text-align: center; ">
     <div style="">
         <strong style="display: block; font-weight-bold; font-size: 18px; margin-bottom: 1rem;">Um diesen Inhalt sehen
             zu können, müssen
@@ -45,14 +52,26 @@
 <script>
 
     function reset() {
-        window.localStorage.clear();
-        document.cookie.split(";").forEach(function (c) {
-            document.cookie = c.replace(/^ +/, "").replace(/=.*/, "=;expires=" + new Date().toUTCString() + ";path=/");
-        });
 
-        if (confirm('Alle Cookies wurden gelöscht und Einstellungen zurückgesetzt, die Seite wird nun neu geladen.')) {
-            parent.window.location.reload();
+
+        if (confirm('Dadurch werden alle Cookies gelöscht und die Seite wird neu geladen, fortfahren?')) {
+            if (confirm('Alle Cookies wurden gelöscht und Einstellungen zurückgesetzt, die Seite wird nun neu geladen.')) {
+                window.localStorage.clear();
+
+                Object.keys(Cookies.get()).forEach(function (cookieName) {
+                    var neededAttributes = {};
+                    Cookies.remove(cookieName, neededAttributes);
+                });
+
+
+                window.top.location.reload();
+
+
+            }
+        } else {
+
         }
+
     }
 </script>
 </html>
